requirements.txt
torch==2.0.1
diffusers
transformers
safetensors
accelerate
Pillow